Prevent precision loss in __DELTA macro
A pointer can fit in an int in 32-bit and 16-bit architectures, but not in 64-bit.

+++ include/tvision/tobjstrm.h
@@ -88,6 +88,7 @@ protected:
 
 #include <dos.h>
 #include <limits.h>
+#include <stddef.h>
 
 const P_id_type P_id_notFound = UINT_MAX;
 
@@ -96,7 +97,7 @@ typedef TStreamable *(*BUILDER)();
 #ifndef __FLAT__
 #define __DELTA( d ) (FP_OFF((TStreamable *)(d *)1)-1)
 #else
-#define __DELTA( d ) ((int)(TStreamable*)(d*)1-1 )
+#define __DELTA( d ) ((ptrdiff_t)(TStreamable*)(d*)1-1 )
 #endif
 
 class TStreamableClass
